config/requirements_base.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile ./config/requirements_base.in
#
asgiref==3.5.2
    # via django
build==0.8.0
    # via pip-tools
click==8.1.3
    # via pip-tools
django==3.2.15
    # via -r ./config/requirements_base.in
packaging==21.3
    # via build
pep517==0.13.0
    # via build
pip-tools==6.8.0
    # via -r ./config/requirements_base.in
pyparsing==3.0.9
    # via packaging
pytz==2022.2.1
    # via django
shellvars-py @ git+https://github.com/birkin/shellvars-py.git
    # via -r ./config/requirements_base.in
sqlparse==0.4.2
    # via django
tomli==2.0.1
    # via
    #   build
    #   pep517
wheel==0.37.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
